Added enable and disable to recordButton and StopRecordingButton also making recordingLabel go back to original text after user hits stop
In this commit I used the property isEnabled and made it true or false depending on if the button was enabled or not. I made isEnabled for stopRecording false in the viewDidLoad function so when the app starts stopRecordingButton will not work. Inside the recordAudio I made recordButton's isEnabled function false to disable it while also enabling the stopRecordingButton to allow the user to stop the recording. In the stopRecording function I then re-Enabled the recordButton and disabled the stopRecordingButton while also resetting the recordingLabel back to what it was at the start thus resetting the application to how it looked when it first opened.